Render code, latex as block nodes when appropriate
Markdown.Code and Markdown.LaTeX nodes should also be rendered as
blocks in other contexts besides Markdown.MD too. Specifically, instead
of only checking for Markdown.MD, we'll dispatch on a Union of all the
block-context nodes when we want to determine block vs. inline based on
the parent node.